Separate sign/verify from swarm.js
