+++ lib/swarm.js
@@ -3,32 +3,13 @@ var subleveldown = require('subleveldown')
 var swarm = require('webrtc-swarm')
 var signalhub = require('signalhub')
 var hyperlog = require('hyperlog')
-var ghsign = require('ghsign')
 var through = require('through2')
 var protobuf = require('protocol-buffers')
 var fs = require('fs')
-var request = require('request')
 
 var messages = protobuf(fs.readFileSync(__dirname + '/../schema.proto'))
 
 module.exports = createSwarm
-
-ghsign = ghsign(function (username, cb) {
-  fs.readFile('./public-keys/' + username + '.keys', 'utf-8', function (_, keys) {
-    if (keys) return cb(null, keys)
-    request('https://github.com/' + username + '.keys', function (_, response) {
-      var keys = response.statusCode === 200 && response.body
-      if (!keys) return cb(new Error('Could not find public keys for ' + username))
-      fs.mkdir('./public-keys', function () {
-        fs.writeFile('./public-keys/' + username + '.keys', keys, function () {
-          cb(null, keys)
-        })
-      })
-    })
-  })
-})
-
-var verifiers = {}
 
 function createSwarm (db, defaultOpts) {
   var emitter = new events.EventEmitter()
@@ -36,11 +17,20 @@ function createSwarm (db, defaultOpts) {
 
   emitter.logs = logs
 
-  var sign
   var processor
+  var sign
+  var verify
 
   emitter.changes = function (name) {
     return logs[name] ? logs[name].changes : 0
+  }
+
+  emitter.sign = function (fn) {
+    sign = fn
+  }
+
+  emitter.verify = function (fn) {
+    verify = fn
   }
 
   emitter.process = function (fn) {
@@ -102,10 +92,6 @@ function createSwarm (db, defaultOpts) {
   }
 
   emitter.send = function (message, cb) {
-    if (!sign && emitter.username) {
-      sign = ghsign.signer(emitter.username)
-    }
-
     var addMessage = function (m, sig) {
       var ch = message.channel || 'friends'
       emitter.addChannel(ch)
@@ -144,14 +130,11 @@ function createSwarm (db, defaultOpts) {
         var val = messages.SignedMessage.decode(data.value)
         var m = messages.Message.decode(val.message)
         var u = m.username
-        var verify = verifiers[u]
-
-        if (u && !verifiers[u]) verifiers[u] = verify = ghsign.verifier(u)
 
         m.change = data.change
 
-        if (verify && val.signature) {
-          verify(val.message, val.signature, function (_, valid) {
+        if (verify) {
+          verify(u, val.message, val.signature, function (_, valid) {
             m.valid = !!valid
             processor(m, cb)
           })
